+++ src/zope/pagetemplate/tests/test_engine.py
@@ -48,6 +48,9 @@ class DummyEngine(object):
     def getTypes(self):
         return {}
 
+    def getCompilerError(self):
+        return SyntaxError
+
 
 class DummyContext(object):
 
@@ -64,20 +67,25 @@ class ZopePythonExprTests(unittest.TestCase):
         expr = ZopePythonExpr('python', 'max(a,b)', DummyEngine())
         self.assertEqual(expr(DummyContext(a=1, b=2)), 2)
 
-    def test_allowed_module_name(self):
+    def test_import_not_possible(self):
         from zope.pagetemplate.engine import ZopePythonExpr
-        expr = ZopePythonExpr('python', '__import__("sys").__name__',
-                              DummyEngine())
-        self.assertEqual(expr(DummyContext()), 'sys')
+        with self.assertRaises(SyntaxError) as err:
+            ZopePythonExpr('python', 'import sys', DummyEngine())
+        if zope.pagetemplate.engine.HAVE_UNTRUSTED:
+            self.assertIn(
+                'SyntaxError: invalid syntax at statement', str(err.exception))
+        else:
+            self.assertEqual(
+                'invalid syntax (<string>, line 1)', str(err.exception))
 
     @unittest.skipUnless(zope.pagetemplate.engine.HAVE_UNTRUSTED,
                          "Needs untrusted")
-    def test_forbidden_module_name(self):
+    def test___import___not_allowed(self):
         from zope.pagetemplate.engine import ZopePythonExpr
-        from zope.security.interfaces import Forbidden
-        expr = ZopePythonExpr('python', '__import__("sys").exit',
-                              DummyEngine())
-        self.assertRaises(Forbidden, expr, DummyContext())
+        with self.assertRaises(SyntaxError) as err:
+            ZopePythonExpr('python', '__import__("sys")', DummyEngine())
+        self.assertIn(
+            '"__import__" is an invalid variable', str(err.exception))
 
     @unittest.skipUnless(zope.pagetemplate.engine.HAVE_UNTRUSTED,
                          "Needs untrusted")