+++ CHANGES.rst
@@ -9,6 +9,9 @@
 
 - Add support for Python 3.9.
 
+- Add support for ``zope.untrustedpython`` on Python 3. With it, Python
+  expressions are now protected. It is activated using the ``untrusted`` extra.
+
 
 4.5.0 (2020-02-10)
 ==================